pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pattern-language-miner"
version = "1.0.0"
description = "A corpus-driven tool for extracting, clustering, and cataloguing reusable content patterns from text, Markdown, and HTML documents."
authors = [{ name = "Pattern Language Miner Contributors" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["nlp", "pattern-language", "documentation", "clustering", "knowledge-graph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "beautifulsoup4>=4.12",
    "click>=8.1",
    "markdown>=3.5",
    "matplotlib>=3.8",
    "networkx>=3.2",
    "nltk==3.7",
    "PyYAML>=6.0",
    "scikit-learn>=1.4",
    "seaborn>=0.13",
    "sentence-transformers>=2.7",
    "tqdm>=4.66",
    "umap-learn>=0.5",
    "weaviate-client>=3.26",
    "jsonschema>=4.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "isort>=5.13",
    "mypy>=1.9",
    "ruff>=0.4",
    "types-PyYAML",
    "types-beautifulsoup4",
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocs-mermaid2-plugin>=1.1",
    "pymdown-extensions>=10.7",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocs-mermaid2-plugin>=1.1",
    "pymdown-extensions>=10.7",
]

[project.scripts]
pattern-miner = "pattern_language_miner.cli:cli"

[project.urls]
Homepage = "https://github.com/your-org/pattern-language-miner"
Documentation = "https://your-org.github.io/pattern-language-miner"
Issues = "https://github.com/your-org/pattern-language-miner/issues"

[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
log_cli = true
log_cli_level = "WARNING"

# ---------------------------------------------------------------------------
# black
# ---------------------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

# ---------------------------------------------------------------------------
# isort
# ---------------------------------------------------------------------------
[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
line_length = 88

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "PT"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
mypy_path = "src"

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/pattern_language_miner"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
